Add Tar DataSource
TODO: Add tests.

I assume that there are only two columns: A string first column, and a
binary second column.

This doesn't quite work for WebDataset format at this time. We
probably would want to have a single string column act as the "base
name", and then use the column names for the suffix of each entry in
the tar file, if we wanted to support that.