WIP: Tar File Format DataSource for Spark (#35)
* Commit of work for final sprint for Round 1 paper submission.

I don't think we need to test this at all, but it is good to have for
posterity.

* Upgrade to spark 3.1.2

Add flac to sox.

* Copy original yamnet in.

See README for source

* Add Tar DataSource

TODO: Add tests.

I assume that there are only two columns: A string first column, and a
binary second column.

This doesn't quite work for WebDataset format at this time. We
probably would want to have a single string column act as the "base
name", and then use the column names for the suffix of each entry in
the tar file, if we wanted to support that.